Reverse polarity of compass bearing rotation
Using the map bearing number directly in the CSS transform rule for the
navigation control results in the compass pointing in the direction of
the rotation, rather than pointing north.

In the core library, some math is applied to the map angle to set the
compass direction (see
https://github.com/mapbox/mapbox-gl-js/blob/d39a7db478a0b686199b750548397d87d61e8f3b/src/ui/control/navigation_control.js#L62).
This is equivalent to reversing the polarity of the bearing, and using
that to set the compass direction, which will now point north.