Remove link to redux-map-gl in documentation
This library hasn't been updated in three years and appears to no longer be actively maintained. It has fallen out of sync with changes made to react-map-gl. For example, the width and height props are still considered separate props and are not part of the viewport state, like they are in react-map-gl.