Fix changing zoom and center at once
Previously, if you wanted to zoom in and pan the map in the same viewport operation, it would get "stuck" at certain latitudes.

If you're zoomed right out such that the equator is at the center of your viewport, and want to zoom in to a point at a much higher zoom level away from the equator, because the center change is handled first, the center change is constrained, so the latitude doesn't change. Then the zoom happens, and you end up at the wrong place.

By changing the order to zoom first, we ensure that the constraining that happens during the center change happens with the correct available center range.